+++ web/secure_screen_wrapper.dart
@@ -1,0 +1,1 @@
+export 'package:secure_screen_wrapper/secure_screen_wrapper_web.dart';